+++ grafana-plugin/e2e-tests/utils/navigation.ts
@@ -1,13 +1,15 @@
-import type { Page, Response } from '@playwright/test';
+import type { Page } from '@playwright/test';
 
 import { BASE_URL } from './constants';
 
 type GrafanaPage = '/plugins/grafana-oncall-app';
 type OnCallPage = 'alert-groups' | 'integrations' | 'escalations' | 'schedules' | 'users';
 
-const _goToPage = (page: Page, url = ''): Promise<Response> => page.goto(`${BASE_URL}${url}`);
+const _goToPage = async (page: Page, url = '') => page.goto(`${BASE_URL}${url}`);
 
-export const goToGrafanaPage = (page: Page, url: GrafanaPage): Promise<Response> => _goToPage(page, url);
+export const goToGrafanaPage = async (page: Page, url: GrafanaPage) => _goToPage(page, url);
 
-export const goToOnCallPage = (page: Page, onCallPage: OnCallPage): Promise<Response> =>
-  _goToPage(page, `/a/grafana-oncall-app/${onCallPage}`);
+export const goToOnCallPage = async (page: Page, onCallPage: OnCallPage) => {
+  await _goToPage(page, `/a/grafana-oncall-app/${onCallPage}`);
+  await page.waitForTimeout(1000);
+};